refactor(runtime/k8s): refactor dp container handling
