+++ pkg/plugins/runtime/k8s/webhooks/injector/injector.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"sort"
 	"strconv"
 	"strings"
 
@@ -12,7 +11,6 @@ import (
 	kube_core "k8s.io/api/core/v1"
 	kube_api "k8s.io/apimachinery/pkg/api/resource"
 	kube_types "k8s.io/apimachinery/pkg/types"
-	kube_intstr "k8s.io/apimachinery/pkg/util/intstr"
 	kube_client "sigs.k8s.io/controller-runtime/pkg/client"
 
 	runtime_k8s "github.com/kumahq/kuma/pkg/config/plugins/runtime/k8s"
@@ -21,6 +19,7 @@ import (
 	core_model "github.com/kumahq/kuma/pkg/core/resources/model"
 	k8s_common "github.com/kumahq/kuma/pkg/plugins/common/k8s"
 	mesh_k8s "github.com/kumahq/kuma/pkg/plugins/resources/k8s/native/api/v1alpha1"
+	"github.com/kumahq/kuma/pkg/plugins/runtime/k8s/containers"
 	"github.com/kumahq/kuma/pkg/plugins/runtime/k8s/metadata"
 	"github.com/kumahq/kuma/pkg/plugins/runtime/k8s/util"
 	tp_k8s "github.com/kumahq/kuma/pkg/transparentproxy/kubernetes"
@@ -35,7 +34,7 @@ var log = core.Log.WithName("injector")
 
 func New(
 	cfg runtime_k8s.Injector,
-	controlPlaneUrl string,
+	controlPlaneURL string,
 	client kube_client.Client,
 	converter k8s_common.Converter,
 ) (*KumaInjector, error) {
@@ -48,20 +47,23 @@ func New(
 		caCert = string(bytes)
 	}
 	return &KumaInjector{
-		cfg:             cfg,
-		controlPlaneUrl: controlPlaneUrl,
-		client:          client,
-		converter:       converter,
-		caCert:          caCert,
+		cfg:       cfg,
+		client:    client,
+		converter: converter,
+		proxyFactory: containers.DataplaneProxyFactory{
+			ControlPlaneURL:    controlPlaneURL,
+			ControlPlaneCACert: caCert,
+			ContainerConfig:    cfg.SidecarContainer.DataplaneContainer,
+			BuiltinDNS:         cfg.BuiltinDNS,
+		},
 	}, nil
 }
 
 type KumaInjector struct {
-	cfg             runtime_k8s.Injector
-	controlPlaneUrl string
-	client          kube_client.Client
-	converter       k8s_common.Converter
-	caCert          string
+	cfg          runtime_k8s.Injector
+	client       kube_client.Client
+	converter    k8s_common.Converter
+	proxyFactory containers.DataplaneProxyFactory
 }
 
 func (i *KumaInjector) InjectKuma(pod *kube_core.Pod) error {
@@ -233,29 +235,11 @@ func (i *KumaInjector) namespaceFor(pod *kube_core.Pod) (*kube_core.Namespace, e
 	return ns, nil
 }
 
-func (i *KumaInjector) proxyConcurrencyFor(annotations map[string]string) (int64, error) {
-	count, ok, err := metadata.Annotations(annotations).GetUint32(metadata.KumaSidecarConcurrencyAnnotation)
-	if ok {
-		return int64(count), err
-	}
-
-	// Note that validation requires the resource limit is not empty.
-	cpuRequest := kube_api.MustParse(i.cfg.SidecarContainer.Resources.Limits.CPU)
-	ncpu := cpuRequest.MilliValue() / 1000
-	if ncpu < 2 {
-		// Only autotune to down to 2 to mitigate the latency
-		// risk if a worker thread blocks.
-		ncpu = 2
-	}
-
-	return ncpu, nil
-}
-
 func (i *KumaInjector) NewSidecarContainer(
 	pod *kube_core.Pod,
 	ns *kube_core.Namespace,
 ) (kube_core.Container, error) {
-	container, err := i.newDpContainer(pod.Annotations, ns)
+	container, err := i.proxyFactory.NewContainer(pod.Annotations, ns)
 	if err != nil {
 		return container, err
 	}
@@ -269,224 +253,6 @@ func (i *KumaInjector) NewSidecarContainer(
 	container.Name = util.KumaSidecarContainerName
 
 	return container, nil
-}
-
-func (i *KumaInjector) NewGatewayContainer(
-	annotations map[string]string,
-	ns *kube_core.Namespace,
-) (kube_core.PodSpec, error) {
-	container, err := i.newDpContainer(annotations, ns)
-	if err != nil {
-		return kube_core.PodSpec{}, err
-	}
-
-	container.Name = util.KumaGatewayContainerName
-
-	return kube_core.PodSpec{
-		Containers: []kube_core.Container{container},
-	}, nil
-}
-
-func (i *KumaInjector) newDpContainer(
-	annotations map[string]string,
-	ns *kube_core.Namespace,
-) (kube_core.Container, error) {
-	mesh := meshName(annotations, ns)
-	env, err := i.sidecarEnvVars(mesh, annotations)
-	if err != nil {
-		return kube_core.Container{}, err
-	}
-
-	cpuCount, err := i.proxyConcurrencyFor(annotations)
-	if err != nil {
-		return kube_core.Container{}, err
-	}
-
-	args := []string{
-		"run",
-		"--log-level=info",
-	}
-
-	if cpuCount > 0 {
-		args = append(args,
-			"--concurrency="+strconv.FormatInt(cpuCount, 10))
-	}
-
-	return kube_core.Container{
-		Image:           i.cfg.SidecarContainer.Image,
-		ImagePullPolicy: kube_core.PullIfNotPresent,
-		Args:            args,
-		Env:             env,
-		SecurityContext: &kube_core.SecurityContext{
-			RunAsUser:  &i.cfg.SidecarContainer.UID,
-			RunAsGroup: &i.cfg.SidecarContainer.GID,
-		},
-		LivenessProbe: &kube_core.Probe{
-			Handler: kube_core.Handler{
-				HTTPGet: &kube_core.HTTPGetAction{
-					Path: "/ready",
-					Port: kube_intstr.IntOrString{
-						IntVal: int32(i.cfg.SidecarContainer.AdminPort),
-					},
-				},
-			},
-			InitialDelaySeconds: i.cfg.SidecarContainer.LivenessProbe.InitialDelaySeconds,
-			TimeoutSeconds:      i.cfg.SidecarContainer.LivenessProbe.TimeoutSeconds,
-			PeriodSeconds:       i.cfg.SidecarContainer.LivenessProbe.PeriodSeconds,
-			SuccessThreshold:    1,
-			FailureThreshold:    i.cfg.SidecarContainer.LivenessProbe.FailureThreshold,
-		},
-		ReadinessProbe: &kube_core.Probe{
-			Handler: kube_core.Handler{
-				HTTPGet: &kube_core.HTTPGetAction{
-					Path: "/ready",
-					Port: kube_intstr.IntOrString{
-						IntVal: int32(i.cfg.SidecarContainer.AdminPort),
-					},
-				},
-			},
-			InitialDelaySeconds: i.cfg.SidecarContainer.ReadinessProbe.InitialDelaySeconds,
-			TimeoutSeconds:      i.cfg.SidecarContainer.ReadinessProbe.TimeoutSeconds,
-			PeriodSeconds:       i.cfg.SidecarContainer.ReadinessProbe.PeriodSeconds,
-			SuccessThreshold:    i.cfg.SidecarContainer.ReadinessProbe.SuccessThreshold,
-			FailureThreshold:    i.cfg.SidecarContainer.ReadinessProbe.FailureThreshold,
-		},
-		Resources: kube_core.ResourceRequirements{
-			Requests: kube_core.ResourceList{
-				kube_core.ResourceCPU:    kube_api.MustParse(i.cfg.SidecarContainer.Resources.Requests.CPU),
-				kube_core.ResourceMemory: kube_api.MustParse(i.cfg.SidecarContainer.Resources.Requests.Memory),
-			},
-			Limits: kube_core.ResourceList{
-				kube_core.ResourceCPU:    kube_api.MustParse(i.cfg.SidecarContainer.Resources.Limits.CPU),
-				kube_core.ResourceMemory: kube_api.MustParse(i.cfg.SidecarContainer.Resources.Limits.Memory),
-			},
-		},
-	}, nil
-}
-
-func (i *KumaInjector) sidecarEnvVars(mesh string, podAnnotations map[string]string) ([]kube_core.EnvVar, error) {
-	envVars := map[string]kube_core.EnvVar{
-		"KUMA_CONTROL_PLANE_URL": {
-			Name:  "KUMA_CONTROL_PLANE_URL",
-			Value: i.controlPlaneUrl,
-		},
-		"KUMA_DATAPLANE_MESH": {
-			Name:  "KUMA_DATAPLANE_MESH",
-			Value: mesh,
-		},
-		"KUMA_DATAPLANE_NAME": {
-			Name: "KUMA_DATAPLANE_NAME",
-			// notice that Pod name might not be available at this time (in case of Deployment, ReplicaSet, etc)
-			// that is why we have to use a runtime reference to POD_NAME instead
-			Value: "$(POD_NAME).$(POD_NAMESPACE)", // variable references get expanded by Kubernetes
-		},
-		"KUMA_DATAPLANE_ADMIN_PORT": {
-			Name:  "KUMA_DATAPLANE_ADMIN_PORT",
-			Value: fmt.Sprintf("%d", i.cfg.SidecarContainer.AdminPort),
-		},
-		"KUMA_DATAPLANE_DRAIN_TIME": {
-			Name:  "KUMA_DATAPLANE_DRAIN_TIME",
-			Value: i.cfg.SidecarContainer.DrainTime.String(),
-		},
-		"KUMA_DATAPLANE_RUNTIME_TOKEN_PATH": {
-			Name:  "KUMA_DATAPLANE_RUNTIME_TOKEN_PATH",
-			Value: "/var/run/secrets/kubernetes.io/serviceaccount/token",
-		},
-		"KUMA_CONTROL_PLANE_CA_CERT": {
-			Name:  "KUMA_CONTROL_PLANE_CA_CERT",
-			Value: i.caCert,
-		},
-	}
-	if i.cfg.BuiltinDNS.Enabled {
-		envVars["KUMA_DNS_ENABLED"] = kube_core.EnvVar{
-			Name:  "KUMA_DNS_ENABLED",
-			Value: "true",
-		}
-
-		envVars["KUMA_DNS_CORE_DNS_PORT"] = kube_core.EnvVar{
-			Name:  "KUMA_DNS_CORE_DNS_PORT",
-			Value: strconv.FormatInt(int64(i.cfg.BuiltinDNS.Port), 10),
-		}
-
-		envVars["KUMA_DNS_CORE_DNS_EMPTY_PORT"] = kube_core.EnvVar{
-			Name:  "KUMA_DNS_CORE_DNS_EMPTY_PORT",
-			Value: strconv.FormatInt(int64(i.cfg.BuiltinDNS.Port+1), 10),
-		}
-
-		envVars["KUMA_DNS_ENVOY_DNS_PORT"] = kube_core.EnvVar{
-			Name:  "KUMA_DNS_ENVOY_DNS_PORT",
-			Value: strconv.FormatInt(int64(i.cfg.BuiltinDNS.Port+2), 10),
-		}
-
-		envVars["KUMA_DNS_CORE_DNS_BINARY_PATH"] = kube_core.EnvVar{
-			Name:  "KUMA_DNS_CORE_DNS_BINARY_PATH",
-			Value: "coredns",
-		}
-	} else {
-		envVars["KUMA_DNS_ENABLED"] = kube_core.EnvVar{
-			Name:  "KUMA_DNS_ENABLED",
-			Value: "false",
-		}
-	}
-
-	// override defaults with cfg env vars
-	for envName, envVal := range i.cfg.SidecarContainer.EnvVars {
-		envVars[envName] = kube_core.EnvVar{
-			Name:  envName,
-			Value: envVal,
-		}
-	}
-
-	// override defaults and cfg env vars with annotations
-	annotationEnvVars, err := metadata.Annotations(podAnnotations).GetMap(metadata.KumaSidecarEnvVarsAnnotation)
-	if err != nil {
-		return nil, err
-	}
-	for envName, envVal := range annotationEnvVars {
-		envVars[envName] = kube_core.EnvVar{
-			Name:  envName,
-			Value: envVal,
-		}
-	}
-
-	var result []kube_core.EnvVar
-	for _, v := range envVars {
-		result = append(result, v)
-	}
-	sort.Stable(EnvVarsByName(result))
-
-	// those values needs to be added before other vars, otherwise expressions like "$(POD_NAME).$(POD_NAMESPACE)" won't be evaluated
-	result = append([]kube_core.EnvVar{
-		{
-			Name: "POD_NAME",
-			ValueFrom: &kube_core.EnvVarSource{
-				FieldRef: &kube_core.ObjectFieldSelector{
-					APIVersion: "v1",
-					FieldPath:  "metadata.name",
-				},
-			},
-		},
-		{
-			Name: "POD_NAMESPACE",
-			ValueFrom: &kube_core.EnvVarSource{
-				FieldRef: &kube_core.ObjectFieldSelector{
-					APIVersion: "v1",
-					FieldPath:  "metadata.namespace",
-				},
-			},
-		},
-		{
-			Name: "INSTANCE_IP",
-			ValueFrom: &kube_core.EnvVarSource{
-				FieldRef: &kube_core.ObjectFieldSelector{
-					APIVersion: "v1",
-					FieldPath:  "status.podIP",
-				},
-			},
-		},
-	}, result...)
-
-	return result, nil
 }
 
 func (i *KumaInjector) NewVolumeMounts(pod *kube_core.Pod) []kube_core.VolumeMount {
@@ -590,12 +356,4 @@ func portsToAnnotationValue(ports []uint32) string {
 		stringPorts[i] = fmt.Sprintf("%d", port)
 	}
 	return strings.Join(stringPorts, ",")
-}
-
-type EnvVarsByName []kube_core.EnvVar
-
-func (a EnvVarsByName) Len() int      { return len(a) }
-func (a EnvVarsByName) Swap(i, j int) { a[i], a[j] = a[j], a[i] }
-func (a EnvVarsByName) Less(i, j int) bool {
-	return a[i].Name < a[j].Name
 }
